+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -41,6 +41,10 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
+   * A callback function to execute when a change is made in the date input. The first parameter is the event. The second parameter is the changed date value.
+   */
+  onChangeRaw: PropTypes.func,
+  /**
    * A callback function to execute when clicking outside of the picker to dismiss it.
    */
   onClickOutside: PropTypes.func,
@@ -71,6 +75,7 @@ const defaultProps = {
   minDate: undefined,
   name: undefined,
   onChange: undefined,
+  onChangeRaw: undefined,
   onClickOutside: undefined,
   onSelect: undefined,
   releaseFocus: undefined,
@@ -96,12 +101,13 @@ class DatePicker extends React.Component {
     };
 
     this.handleChange = this.handleChange.bind(this);
+    this.handleChangeRaw = this.handleChangeRaw.bind(this);
     this.handleOnSelect = this.handleOnSelect.bind(this);
     this.handleOnClickOutside = this.handleOnClickOutside.bind(this);
   }
 
   handleOnSelect(selectedDate, event) {
-    this.requestFocus();
+    this.releaseFocus();
 
     if (this.props.onSelect) {
       this.props.onSelect(event, selectedDate ? selectedDate.format() : '');
@@ -109,14 +115,14 @@ class DatePicker extends React.Component {
   }
 
   handleOnClickOutside(event) {
-    this.requestFocus();
+    this.releaseFocus();
 
     if (this.props.onClickOutside) {
       this.props.onClickOutside(event);
     }
   }
 
-  requestFocus() {
+  releaseFocus() {
     // The picker will be dismissed and the focus will be released so that the containing component (e.g. modal) can regain focus.
     if (this.props.releaseFocus) {
       this.props.releaseFocus();
@@ -129,7 +135,13 @@ class DatePicker extends React.Component {
     });
 
     if (this.props.onChange) {
-      this.props.onChange(event, date ? date.format() : '');
+      this.props.onChange(event, date && date.isValid() ? date.format() : '');
+    }
+  }
+
+  handleChangeRaw(event) {
+    if (this.props.onChangeRaw) {
+      this.props.onChangeRaw(event, event.target.value);
     }
   }
 
@@ -143,6 +155,7 @@ class DatePicker extends React.Component {
       minDate,
       name,
       onChange,
+      onChangeRaw,
       onClickOutside,
       onSelect,
       requestFocus,
@@ -164,6 +177,7 @@ class DatePicker extends React.Component {
         {...customProps}
         selected={this.state.selectedDate}
         onChange={this.handleChange}
+        onChangeRaw={this.handleChangeRaw}
         onClickOutside={this.handleOnClickOutside}
         onSelect={this.handleOnSelect}
         customInput={<DateInput inputAttributes={inputAttributes} releaseFocus={releaseFocus} requestFocus={requestFocus} />}
@@ -190,6 +204,7 @@ class DatePicker extends React.Component {
         {...customProps}
         selected={this.state.selectedDate}
         onChange={this.handleChange}
+        onChangeRaw={this.handleChangeRaw}
         onClickOutside={this.handleOnClickOutside}
         onSelect={this.handleOnSelect}
         customInput={<DateInput inputAttributes={inputAttributes} releaseFocus={releaseFocus} requestFocus={requestFocus} />}