+++ packages/terra-site/src/examples/date-time-picker/DateTimePickerDefault.jsx
@@ -1,11 +1,8 @@
 import React from 'react';
-import moment from 'moment';
 import DateTimePickerExampleTemplate from './DateTimePickerExampleTemplate';
 
 const DateTimePickerExample = () => (
-  <DateTimePickerExampleTemplate
-    value={moment().format()}
-  />
+  <DateTimePickerExampleTemplate />
 );
 
 export default DateTimePickerExample;