+++ packages/terra-site/src/examples/date-time-picker/DateTimePickerFilterDates.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
-import moment from 'moment';
 import DateTimePickerExampleTemplate from './DateTimePickerExampleTemplate';
 
+const isWeekday = (date) => {
+  const day = date.day();
+  return day !== 0 && day !== 6;
+};
+
 const DateTimePickerExample = () => (
-  <DateTimePickerExampleTemplate
-    value={moment().format()}
-  />
+  <DateTimePickerExampleTemplate filterDate={isWeekday} />
 );
 
 export default DateTimePickerExample;