+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added onChangeRaw, onClickOutside and onSelect props.
 
 1.5.0 - (August 8, 2017)
 ------------------